+++ common-core/src/test/java/cn/featherfly/common/lang/StringFormatterTest.java
@@ -8,6 +8,7 @@ import java.util.Map;
 import org.testng.annotations.Test;
 
 import cn.featherfly.common.lang.string.StringFormatter;
+import cn.featherfly.common.lang.vo.User;
 import cn.featherfly.common.structure.HashChainMap;
 
 /**
@@ -20,11 +21,14 @@ import cn.featherfly.common.structure.HashChainMap;
 public class StringFormatterTest {
 
     String expected = "hello yufei at 2020 from [yi] at 12:15";
+    String expectedUser = "hello featherfly , age is 20";
     String name = "yufei";
     String name2 = "yi";
     int year = 2020;
     String time = "12:15";
     String actual = null;
+
+    User user = new User("featherfly", 20);
 
     Map<String, Object> argsMap = new HashChainMap<String, Object>().putChain("name", name).putChain("year", year)
             .putChain("name2", name2).putChain("time", time);
@@ -37,6 +41,9 @@ public class StringFormatterTest {
 
         actual = formatter.format("hello {name} at {year} from [{name2}] at {time}", argsMap);
         assertEquals(actual, expected);
+
+        actual = formatter.format("hello {name} , age is {age}", user);
+        assertEquals(actual, expectedUser);
 
         formatter = new StringFormatter('$', '$');